Modelまで終了

+++ DirectX12CG/Engin/Camera/DebugCamera.cpp
@@ -9,7 +9,7 @@ using namespace DirectX;
 void MCB::DebugCamera::Inilialize()
 {
 	view_.CreateMatrixView(XMFLOAT3(0.0f, 0.0f, -10.0f), XMFLOAT3(0.0f, 0.0f, 0.0f), XMFLOAT3(0.0f, 1.0f, 0.0f));
-	projection_.CreateMatrixProjection(XMConvertToRadians(45.0f), (float)DxWindow::GetInstance()->sWINDOW_WIDTH_ / DxWindow::GetInstance()->sWINDOW_HEIGHT_, 0.1f, 4000.0f);
+	projection_.CreateMatrixProjection(XMConvertToRadians(45.0f), static_cast<float>(DxWindow::GetInstance()->sWINDOW_WIDTH_ / DxWindow::GetInstance()->sWINDOW_HEIGHT_), 0.1f, 4000.0f);
 
 	eye_.vec_.x_ = view_.eye_.x;
 	eye_.vec_.y_ = view_.eye_.y;
@@ -61,7 +61,7 @@ void MCB::DebugCamera::Update()
 
 #pragma region ’Ž‹“_‚ÆŽ‹“_‚Ì‹——£•ÏX
 	if (input_->IsKeyDown(DIK_RSHIFT)) {
-		disEyeTarget_ += -input_->GetMousePosition().z * (disEyeTarget_ * 0.001f);
+		disEyeTarget_ += static_cast<float>( -input_->GetMousePosition().z) * (disEyeTarget_ * 0.001f);
 	}
 	if (disEyeTarget_ < 10) {
 		disEyeTarget_ = 10;
